Added options and initted testing for fields
