+++ extract/extract.go
@@ -22,9 +22,11 @@ func ParseDelimitted(fullPath string, delimiter string, fieldMap FieldMap) ([]Ro
 
 	// set up buffer for reading
 	csvReader := csv.NewReader(f_in)
+	
 	inScanner := bufio.NewScanner(f_in)
 
 	for inScanner.Scan() {
+		csvReader.ReadAll()
 
 	}
